AAP: Define and use clz/ctz for fixed width types
Since the builtin clz/ctz functions are defined based on the basic
integer types, using these functions when operating based on a fixed
width will not operate as expected on all targets.

This change defines macros that expand to the correct function, and
also ensures that types used in functions are used correctly based on
their fixed width where appropriate as opposed to the basic integer
types.