AAP: Ensure shift of exponent does not exceed type width
In floatdisf, the exponent was stored in an int type before being
shifted into position, however if int is 16 bits only then the shift
amount is larger than the type width. This change casts the exponent to
the unsigned 32-bit type to ensure that the shift can occur without
losing bits.